requirement.txt
python-pptx
Pillow
PyQt5
